+++ tests/test_lan.py
@@ -8,28 +8,26 @@ from simpletr64.devicetr64 import DeviceTR64
 class TestLan(unittest.TestCase):
 
     def test_AmountOfHostsConnected(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Lan(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        lan = Lan(box)
-        amount = lan.getAmountOfHostsConnected()
+        amount = box.getAmountOfHostsConnected()
 
         self.assertTrue(amount > 0)
 
     def test_HostDetailsByIndex(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Lan(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        lan = Lan(box)
-        hostDetails = lan.getHostDetailsByIndex(0)
+        hostDetails = box.getHostDetailsByIndex(0)
 
         self.assertTrue(hostDetails.ipaddress)
         self.assertTrue(hostDetails.hostname)
@@ -41,17 +39,16 @@ class TestLan(unittest.TestCase):
         self.assertTrue(len(hostDetails.raw.keys()) > 0)
 
     def test_HostDetailsByMac(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Lan(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        lan = Lan(box)
-        hostDetailsForFirst = lan.getHostDetailsByIndex(0)
+        hostDetailsForFirst = box.getHostDetailsByIndex(0)
 
-        hostDetails = lan.getHostDetailsByMACAddress(hostDetailsForFirst.macAddress)
+        hostDetails = box.getHostDetailsByMACAddress(hostDetailsForFirst.macAddress)
 
         self.assertTrue(hostDetails.ipaddress)
         self.assertTrue(hostDetails.hostname)
@@ -63,15 +60,14 @@ class TestLan(unittest.TestCase):
         self.assertTrue(len(hostDetails.raw.keys()) > 0)
 
     def test_EhternetInfo(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Lan(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        lan = Lan(box)
-        ethernetInfo = lan.getEthernetInfo()
+        ethernetInfo = box.getEthernetInfo()
 
         self.assertTrue(ethernetInfo.enabled or not ethernetInfo.enabled)
         self.assertTrue(ethernetInfo.status)
@@ -81,15 +77,14 @@ class TestLan(unittest.TestCase):
         self.assertTrue(len(ethernetInfo.raw.keys()) > 0)
 
     def test_EhternetStatistic(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Lan(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        lan = Lan(box)
-        ethernetStatistic = lan.getEthernetStatistic()
+        ethernetStatistic = box.getEthernetStatistic()
 
         self.assertTrue(ethernetStatistic.bytesSent >= 0)
         self.assertTrue(ethernetStatistic.bytesReceived >= 0)