+++ tests/test_wifi.py
@@ -8,15 +8,14 @@ from simpletr64.devicetr64 import DeviceTR64
 class TestWifi(unittest.TestCase):
 
     def test_WifiInfo(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Wifi(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        wifi = Wifi(box)
-        wifiInfo = wifi.getWifiInfo(1)
+        wifiInfo = box.getWifiInfo(1)
 
         self.assertTrue(wifiInfo.enabled or not wifiInfo.enabled)
         self.assertTrue(wifiInfo.status)
@@ -31,47 +30,44 @@ class TestWifi(unittest.TestCase):
         self.assertTrue(len(wifiInfo.raw.keys()) > 0)
 
     def test_WifiStatistic(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Wifi(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        wifi = Wifi(box)
-        stats = wifi.getStatistic(1)
+        stats = box.getStatistic(1)
 
         self.assertTrue(stats[0] >= 0)
         self.assertTrue(stats[1] >= 0)
 
     def test_WifiPacketStatistic(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Wifi(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        wifi = Wifi(box)
-        stats = wifi.getPacketStatistic(1)
+        stats = box.getPacketStatistic(1)
 
         self.assertTrue(stats[0] >= 0)
         self.assertTrue(stats[1] >= 0)
 
     def test_WifiDeviceInfo(self):
-        box = DeviceTR64(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
+        box = Wifi(hostname=defaults.test_host, port=defaults.test_port, protocol=defaults.test_protocol)
         box.setupTR64Device("fritz.box")
         box.username = defaults.test_user
         box.password = defaults.test_pw
         box.httpProxy = defaults.test_httpProxy
         box.httpsProxy = defaults.test_httpsProxy
 
-        wifi = Wifi(box)
-        amount = wifi.getTotalAssociations(1)
+        amount = box.getTotalAssociations(1)
         self.assertTrue(amount > 0, "Amount " + str(amount))
 
         # get first device in the list on the first wifi interface
-        deviceInfo = wifi.getGenericAssociatedDeviceInfo(0 ,1)
+        deviceInfo = box.getGenericAssociatedDeviceInfo(0 ,1)
 
         self.assertTrue(deviceInfo.macAddress)
         self.assertTrue(deviceInfo.ipAddress)