+++ HISTORY.rst
@@ -11,6 +11,7 @@ Release History
 * Improved error handling on XML parsing problems of any content
 * Changed the declaration of getGenericAssociatedDeviceInfo for consistency reasons
 * Changed for all Wifi actions that the interface id is now optional
+* Changed all TR64 action classes to a sub class of DeviceTR64
 * Fix an issue with relative URL's in the device definition
 
 1.0.1 (2016-01-04)